backend/requirements.txt
# Core
Django>=5.0
djangorestframework>=3.14
django-cors-headers>=4.3.1
requests>=2.31
python-dotenv>=1.0.1

# Production (comment out for dev if not using PostgreSQL / S3)
psycopg2-binary>=2.9
# django-storages>=1.14
# boto3>=1.34

# Optional: MARC/cataloging metadata lookup
# PyZ3950>=2.0

# Deployment
gunicorn>=21.2
